+++ src/EntityAliasDependencyProviderInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Drupal\wmpathauto;
+
+use Drupal\Component\Plugin\PluginInspectionInterface;
+use Drupal\Core\Entity\EntityInterface;
+
+interface EntityAliasDependencyProviderInterface extends PluginInspectionInterface
+{
+    public function addDependencies(EntityAliasDependencyCollectionInterface $dependencies, EntityInterface $entity): void;
+}