+++ src/PatternTokenDependencyProviderInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\wmpathauto;
+
+use Drupal\Component\Plugin\PluginInspectionInterface;
+
+interface PatternTokenDependencyProviderInterface extends PluginInspectionInterface
+{
+    public function addDependencies(array $tokens, array $data, array $options, EntityAliasDependencyCollectionInterface $dependencies): void;
+}